261- intro to redux(adding notes)
Key points:
(7:56) All global states live in a centralised container; It is the single source of truth of global state in the app.
The store is also the place where one or multiple reducers live.

Each reducer is a pure function that has a single task to calculate the next state.

### Redux Cycle Goal
To make the state update logic separate from the rest of the application