267- Supplement useSelector notes
index.js
```
root.render(
  <React.StrictMode>
    <Provider store={store}>
      <App />
    </Provider>
  </React.StrictMode>
);
```

The useSelector hook takes a callback function, and this function takes a single argument of the entire Redux store, and from that store, we can simply get the data we want.

```
  const customer = useSelector((store) => store.customer.fullName);
```

The store.customer comes from here:
in store.js
```
  const rootReducer = combineReducers({
    account: accountReducer,
    **customer**: customerReducer,
  })
```

(4:57) ** The useSelector hook provided by "react-redux" basically creates a subscription to the store.**
Whenever the store changes, the component that subscribes to that store will re-render. (When the state in store changes, the component re-renders)